+++ src/m365/spfx/commands/project/project-model/YoRcJson.ts
@@ -1,7 +1,7 @@
 import { JsonFile } from ".";
 
 export interface YoRcJson extends JsonFile {
-  "@microsoft/generator-sharepoint": {
+  "@microsoft/generator-sharepoint"?: {
     componentType?: string;
     environment?: string;
     framework?: string;
@@ -9,6 +9,10 @@ export interface YoRcJson extends JsonFile {
     isDomainIsolated?: boolean;
     nodeVersion?: string;
     packageManager?: string;
+    sdkVersions?: {
+      "@microsoft/microsoft-graph-client"?: string;
+      "@microsoft/teams-js"?: string;
+    };
     template?: string;
     version?: string;
   }