+++ src/m365/spfx/commands/project/project-upgrade/rules/FN010003_YORC_packageManager.spec.ts
@@ -33,6 +33,15 @@ describe('FN010003_YORC_packageManager', () => {
     assert.strictEqual(findings.length, 0);
   });
 
+  it('returns notification if @microsoft/generator-sharepoint is not set', () => {
+    const project: Project = {
+      path: '/usr/tmp',
+      yoRcJson: {}
+    };
+    rule.visit(project, findings);
+    assert.strictEqual(findings.length, 1);
+  });
+
   it('returns notification if packageManager is not up-to-date', () => {
     const project: Project = {
       path: '/usr/tmp',