+++ src/m365/spo/commands/field/field-set.ts
@@ -1,9 +1,11 @@
+import { AxiosRequestConfig } from 'axios';
 import { Logger } from '../../../../cli/Logger';
 import config from '../../../../config';
 import GlobalOptions from '../../../../GlobalOptions';
 import request from '../../../../request';
 import { formatting } from '../../../../utils/formatting';
 import { ClientSvcResponse, ClientSvcResponseContents, spo } from '../../../../utils/spo';
+import { urlUtil } from '../../../../utils/urlUtil';
 import { validation } from '../../../../utils/validation';
 import SpoCommand from '../../../base/SpoCommand';
 import commands from '../../commands';
@@ -18,6 +20,7 @@ interface Options extends GlobalOptions {
   name?: string;
   listId?: string;
   listTitle?: string;
+  listUrl?: string;
   updateExistingLists?: boolean;
   webUrl: string;
 }
@@ -48,6 +51,7 @@ class SpoFieldSetCommand extends SpoCommand {
         name: typeof args.options.name !== 'undefined',
         listId: typeof args.options.listId !== 'undefined',
         listTitle: typeof args.options.listTitle !== 'undefined',
+        listUrl: typeof args.options.listUrl !== 'undefined',
         updateExistingLists: !!args.options.updateExistingLists
       });
     });
@@ -63,6 +67,9 @@ class SpoFieldSetCommand extends SpoCommand {
       },
       {
         option: '--listTitle [listTitle]'
+      },
+      {
+        option: '--listUrl [listUrl]'
       },
       {
         option: '-i, --id [id]'
@@ -83,31 +90,32 @@ class SpoFieldSetCommand extends SpoCommand {
     this.validators.push(
       async (args: CommandArgs) => {
         const isValidSharePointUrl: boolean | string = validation.isValidSharePointUrl(args.options.webUrl);
-	    if (isValidSharePointUrl !== true) {
-	      return isValidSharePointUrl;
-	    }
+        if (isValidSharePointUrl !== true) {
+          return isValidSharePointUrl;
+        }
 
-	    if (args.options.listId && args.options.listTitle) {
-	      return `Specify listId or listTitle but not both`;
-	    }
+        const listOptions: any[] = [args.options.listId, args.options.listTitle, args.options.listUrl];
+        if (listOptions.some(item => item !== undefined) && listOptions.filter(item => item !== undefined).length > 1) {
+          return `Specify either list id or title or list url, but not multiple`;
+        }
 
-	    if (args.options.listId &&
-	      !validation.isValidGuid(args.options.listId)) {
-	      return `${args.options.listId} in option listId is not a valid GUID`;
-	    }
+        if (args.options.listId &&
+          !validation.isValidGuid(args.options.listId)) {
+          return `${args.options.listId} in option listId is not a valid GUID`;
+        }
 
-	    if (args.options.id &&
-	      !validation.isValidGuid(args.options.id)) {
-	      return `${args.options.id} in option id is not a valid GUID`;
-	    }
+        if (args.options.id &&
+          !validation.isValidGuid(args.options.id)) {
+          return `${args.options.id} in option id is not a valid GUID`;
+        }
 
-	    return true;
+        return true;
       }
     );
   }
 
   #initOptionSets(): void {
-  	this.optionSets.push(['id', 'title', 'name']);
+    this.optionSets.push(['id', 'title', 'name']);
   }
 
   public allowUnknownOptions(): boolean | undefined {
@@ -123,27 +131,30 @@ class SpoFieldSetCommand extends SpoCommand {
       const reqDigest = await spo.getRequestDigest(args.options.webUrl);
       const requestDigest = reqDigest.FormDigestValue;
 
-      let list = undefined;
-      if (args.options.listId || args.options.listTitle) {
-        const listQuery: string = args.options.listId ?
-          `<Method Id="663" ParentId="7" Name="GetById"><Parameters><Parameter Type="Guid">${formatting.escapeXml(args.options.listId)}</Parameter></Parameters></Method>` :
-          `<Method Id="663" ParentId="7" Name="GetByTitle"><Parameters><Parameter Type="String">${formatting.escapeXml(args.options.listTitle)}</Parameter></Parameters></Method>`;
+      let fieldsParentIdentity = '<Property Id="5" ParentId="3" Name="Web" /><StaticProperty Id="3" TypeId="{3747adcd-a3c3-41b9-bfab-4a64dd2f1e0a}" Name="Current" />';
 
-        const requestOptions: any = {
+      if (args.options.listId || args.options.listTitle || args.options.listUrl) {
+        let requestData = '';
+        if (args.options.listId) {
+          requestData = `<Request AddExpandoFieldTypeSuffix="true" SchemaVersion="15.0.0.0" LibraryVersion="16.0.0.0" ApplicationName="${config.applicationName}" xmlns="http://schemas.microsoft.com/sharepoint/clientquery/2009"><Actions><ObjectPath Id="664" ObjectPathId="663" /><Query Id="665" ObjectPathId="663"><Query SelectAllProperties="false"><Properties /></Query></Query></Actions><ObjectPaths><Method Id="663" ParentId="7" Name="GetById"><Parameters><Parameter Type="Guid">${formatting.escapeXml(args.options.listId)}</Parameter></Parameters></Method><Property Id="7" ParentId="5" Name="Lists" /><Property Id="5" ParentId="3" Name="Web" /><StaticProperty Id="3" TypeId="{3747adcd-a3c3-41b9-bfab-4a64dd2f1e0a}" Name="Current" /></ObjectPaths></Request>`;
+        }
+        else if (args.options.listTitle) {
+          requestData = `<Request AddExpandoFieldTypeSuffix="true" SchemaVersion="15.0.0.0" LibraryVersion="16.0.0.0" ApplicationName="${config.applicationName}" xmlns="http://schemas.microsoft.com/sharepoint/clientquery/2009"><Actions><ObjectPath Id="664" ObjectPathId="663" /><Query Id="665" ObjectPathId="663"><Query SelectAllProperties="false"><Properties /></Query></Query></Actions><ObjectPaths><Method Id="663" ParentId="7" Name="GetByTitle"><Parameters><Parameter Type="String">${formatting.escapeXml(args.options.listTitle)}</Parameter></Parameters></Method><Property Id="7" ParentId="5" Name="Lists" /><Property Id="5" ParentId="3" Name="Web" /><StaticProperty Id="3" TypeId="{3747adcd-a3c3-41b9-bfab-4a64dd2f1e0a}" Name="Current" /></ObjectPaths></Request>`;
+        }
+        else if (args.options.listUrl) {
+          const listServerRelativeUrl: string = urlUtil.getServerRelativePath(args.options.webUrl, args.options.listUrl);
+          requestData = `<Request AddExpandoFieldTypeSuffix="true" SchemaVersion="15.0.0.0" LibraryVersion="16.0.0.0" ApplicationName="${config.applicationName}" xmlns="http://schemas.microsoft.com/sharepoint/clientquery/2009"><Actions><ObjectPath Id="2" ObjectPathId="1" /><ObjectPath Id="4" ObjectPathId="3" /><ObjectPath Id="6" ObjectPathId="5" /><Query Id="7" ObjectPathId="5"><Query SelectAllProperties="true"><Properties /></Query></Query></Actions><ObjectPaths><StaticProperty Id="1" TypeId="{3747adcd-a3c3-41b9-bfab-4a64dd2f1e0a}" Name="Current" /><Property Id="3" ParentId="1" Name="Web" /><Method Id="5" ParentId="3" Name="GetList"><Parameters><Parameter Type="String">${listServerRelativeUrl}</Parameter></Parameters></Method></ObjectPaths></Request>`;
+        }
+
+        const requestOptions: AxiosRequestConfig = {
           url: `${args.options.webUrl}/_vti_bin/client.svc/ProcessQuery`,
           headers: {
             'X-RequestDigest': requestDigest
           },
-          data: `<Request AddExpandoFieldTypeSuffix="true" SchemaVersion="15.0.0.0" LibraryVersion="16.0.0.0" ApplicationName="${config.applicationName}" xmlns="http://schemas.microsoft.com/sharepoint/clientquery/2009"><Actions><ObjectPath Id="664" ObjectPathId="663" /><Query Id="665" ObjectPathId="663"><Query SelectAllProperties="false"><Properties /></Query></Query></Actions><ObjectPaths>${listQuery}<Property Id="7" ParentId="5" Name="Lists" /><Property Id="5" ParentId="3" Name="Web" /><StaticProperty Id="3" TypeId="{3747adcd-a3c3-41b9-bfab-4a64dd2f1e0a}" Name="Current" /></ObjectPaths></Request>`
+          data: requestData
         };
 
-        list = await request.post<string>(requestOptions);
-      }
-
-      // by default retrieve the column from the site
-      let fieldsParentIdentity: string = '<Property Id="5" ParentId="3" Name="Web" /><StaticProperty Id="3" TypeId="{3747adcd-a3c3-41b9-bfab-4a64dd2f1e0a}" Name="Current" />';
-
-      if (list) {
+        const list = await request.post<string>(requestOptions);
         const json: ClientSvcResponse = JSON.parse(list);
         const response: ClientSvcResponseContents = json[0];
         if (response.ErrorInfo) {
@@ -159,7 +170,7 @@ class SpoFieldSetCommand extends SpoCommand {
         `<Method Id="663" ParentId="7" Name="GetById"><Parameters><Parameter Type="Guid">${formatting.escapeXml(args.options.id)}</Parameter></Parameters></Method>` :
         `<Method Id="663" ParentId="7" Name="GetByInternalNameOrTitle"><Parameters><Parameter Type="String">${formatting.escapeXml(args.options.name || args.options.title)}</Parameter></Parameters></Method>`;
 
-      let requestOptions: any = {
+      let requestOptions: AxiosRequestConfig = {
         url: `${args.options.webUrl}/_vti_bin/client.svc/ProcessQuery`,
         headers: {
           'X-RequestDigest': requestDigest
@@ -202,6 +213,7 @@ class SpoFieldSetCommand extends SpoCommand {
       'webUrl',
       'listId',
       'listTitle',
+      'listUrl',
       'id',
       'name',
       'title',