+++ src/m365/spo/commands/listitem/listitem-roleinheritance-reset.ts
@@ -1,7 +1,9 @@
+import { AxiosRequestConfig } from 'axios';
 import { Logger } from '../../../../cli/Logger';
 import GlobalOptions from '../../../../GlobalOptions';
 import request from '../../../../request';
 import { formatting } from '../../../../utils/formatting';
+import { urlUtil } from '../../../../utils/urlUtil';
 import { validation } from '../../../../utils/validation';
 import SpoCommand from '../../../base/SpoCommand';
 import commands from '../../commands';
@@ -15,6 +17,7 @@ interface Options extends GlobalOptions {
   listItemId: number;
   listId?: string;
   listTitle?: string;
+  listUrl?: string;
 }
 
 class SpoListItemRoleInheritanceResetCommand extends SpoCommand {
@@ -39,7 +42,8 @@ class SpoListItemRoleInheritanceResetCommand extends SpoCommand {
     this.telemetry.push((args: CommandArgs) => {
       Object.assign(this.telemetryProperties, {
         listId: typeof args.options.listId !== 'undefined',
-        listTitle: typeof args.options.listTitle !== 'undefined'
+        listTitle: typeof args.options.listTitle !== 'undefined',
+        listUrl: typeof args.options.listUrl !== 'undefined'
       });
     });
   }
@@ -57,6 +61,9 @@ class SpoListItemRoleInheritanceResetCommand extends SpoCommand {
       },
       {
         option: '--listTitle [listTitle]'
+      },
+      {
+        option: '--listUrl [listUrl]'
       }
     );
   }
@@ -83,20 +90,24 @@ class SpoListItemRoleInheritanceResetCommand extends SpoCommand {
   }
 
   #initOptionSets(): void {
-    this.optionSets.push(['listId', 'listTitle']);
+    this.optionSets.push(['listId', 'listTitle', 'listUrl']);
   }
 
   public async commandAction(logger: Logger, args: CommandArgs): Promise<void> {
-    let requestUrl: string = `${args.options.webUrl}/_api/web/lists`;
+    let requestUrl: string = `${args.options.webUrl}/_api/web`;
 
     if (args.options.listId) {
-      requestUrl += `(guid'${formatting.encodeQueryParameter(args.options.listId)}')`;
+      requestUrl += `/lists(guid'${formatting.encodeQueryParameter(args.options.listId)}')`;
     }
-    else {
-      requestUrl += `/getbytitle('${formatting.encodeQueryParameter(args.options.listTitle as string)}')`;
+    else if (args.options.listTitle) {
+      requestUrl += `/lists/getbytitle('${formatting.encodeQueryParameter(args.options.listTitle)}')`;
+    }
+    else if (args.options.listUrl) {
+      const listServerRelativeUrl: string = urlUtil.getServerRelativePath(args.options.webUrl, args.options.listUrl);
+      requestUrl += `/GetList('${formatting.encodeQueryParameter(listServerRelativeUrl)}')`;
     }
 
-    const requestOptions: any = {
+    const requestOptions: AxiosRequestConfig = {
       url: `${requestUrl}/items(${args.options.listItemId})/resetroleinheritance`,
       method: 'POST',
       headers: {